DeviceStore/catalogNavUpdate0.1
