+++ src/App.js
@@ -1,21 +1,175 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
 import MainPage from "./pages/MainPage/MainPage";
-
+import CatalogPage from "./pages/CatalogPage/CatalogPage";
+import navigationPc from "./resources/png/NavigationPC.png";
+import navigationConsole from "./resources/png/NavigationConsole.png";
+import navigationLaptop from "./resources/png/NavigationLaptop.png";
+import navigationHeadPhone from "./resources/png/NavigationHeadPhone.png";
+import navigationGadgets from "./resources/png/NavigationGadgets.png";
+import arrowCatalog from "./resources/svg/arrowCatalog.svg";
+import desktop from "./resources/svg/desktopBurger.svg";
+import gamepad from "./resources/svg/gamepadBurger.svg";
+import laptop from "./resources/svg/laptopBurger.svg";
+import mouse from "./resources/svg/mouseBurger.svg";
+import disk from "./resources/svg/diskBurger.svg";
+import promotion from "./resources/svg/promotionBurger.svg";
+import sellers from "./resources/svg/sellersBurger.svg";
+import Header from "./Components/Header/Header";
+import { Footer } from "./Components/Footer/Footer";
+import Catalog from "./Components/Catalog/Catalog";
+const modalInfo = [
+  {
+    firstList: [
+      "Готові рішення",
+      "Процесори",
+      "Монітори",
+      "Кулери",
+      "Материнські плати",
+      "Відеокарти",
+      "Оперативна пам’ять",
+    ],
+    secondList: [
+      "SSD накопичувачі",
+      "Жорсткі диски",
+      "Блоки живлення",
+      "Вентилятори",
+      "Корпуси плати",
+      "Всі ПК та комплектуючі",
+    ],
+    img: navigationPc,
+    tabsImg: desktop,
+    arrowCatalog,
+    catalog: "ПК та комплектуючі",
+    id: "Computers"
+  },
+  {
+    firstList: [],
+    secondList: [
+      "Ігрові консолі",
+      "Аксесуари для консолей",
+      "Ігрові маніпулятори",
+      "Аксесуари для приставок",
+      "Ігри",
+      "Зберігання та носії для ігор",
+      "Всі консолі та аксесуари",
+    ],
+    img: navigationConsole,
+    tabsImg: gamepad,
+    arrowCatalog,
+    catalog: "Консолі та аксесуари",
+    id: "Consoles"
+  },
+  {
+    firstList: [],
+    secondList: [
+      "Ноутбуки",
+      "Аксесуари для ноутбука",
+      "Комплектуючи",
+      "USB-хаби та картріди",
+      "Зовнішні SSD",
+      "Всі ноутбуки та комплектуючі",
+    ],
+    img: navigationLaptop,
+    tabsImg: laptop,
+    arrowCatalog,
+    catalog: "Ноутбуки",
+    id: "Laptop"
+  },
+  {
+    firstList: [
+      "Веб-камери",
+      "Мишки",
+      "Клавіатури",
+      "Гарнітури та аудіо",
+      "Комплекти",
+      "Геймпади, джойстики та керма",
+    ],
+    secondList: [
+      "Килимки для мишок",
+      "Аксесуари для клавіатури",
+      "Окуляри та шоломи VR",
+      "Стрімерське обладнання",
+      "Столи та стільці",
+      "Вся переферія",
+    ],
+    img: navigationHeadPhone,
+    tabsImg: mouse,
+    arrowCatalog,
+    catalog: "Переферія",
+    id: "Periphery"
+  },
+  {
+    firstList: [],
+    secondList: [
+      "Аксесуари для робочого місця",
+      "Накопичувачі пам’яті",
+      "Кабелі та переходники",
+      "Зарядні пристрої",
+      "Джерела живлення",
+      "Зовнішні та внутрішні кишені",
+      "Брендові товари",
+      "Всі аксесуари",
+    ],
+    img: navigationGadgets,
+    tabsImg: disk,
+    arrowCatalog,
+    catalog: "Аксесуари",
+    id: 'Accessories'
+  },
+  {
+    firstList: [],
+    secondList: [
+      "Ноутбуки",
+      "Аксесуари для ноутбука",
+      "Комплектуючи",
+      "USB-хаби та картріди",
+      "Зовнішні SSD",
+      "Всі ноутбуки та комплектуючі",
+    ],
+    img: navigationLaptop,
+    tabsImg: promotion,
+    arrowCatalog,
+    catalog: "Акції",
+      firstList: [],
+      secondList: [],
+      tabsImg: promotion,
+      catalog: "Акції",
+  },
+  {
+    firstList: [
+      "Про нас",
+      "FAQ та допомога",
+      "Гарантійне обслуговування",
+      "Доставка і оплата",
+      "Обмін та повернення",
+      "Політика конфіденційності",
+    ],
+    secondList: [],
+    tabsImg: sellers,
+    arrowCatalog,
+    catalog: "Покупцеві",
+  },
+];
 const App = () => {
   return (
-    <Routes>
-      <Route path="/" element={<MainPage />} />
-      <Route path="/product/:id" />
-      <Route path="/catalog" element={<MainPage />} />
-      <Route path="/catalog/:category" />
-      <Route path="/catalog/:category/:subcategory" />
-      <Route path="/shop" />
-      <Route path="/user-agreement" />
-      <Route path="/cart" />
-      <Route path="/order" />
-      <Route path="*" />
-    </Routes>
+    <>
+      <Header modalInfo={modalInfo} />
+      <Catalog modalInfo={modalInfo} />
+      <Routes>
+        <Route path="/" element={<MainPage/>} />
+        <Route path="/product/:id" />
+        <Route path="/catalog" element={<MainPage/>} />
+        <Route path="/catalog/:category" element={<CatalogPage/>} />
+        <Route path="/catalog/:category/:subcategory" element={<CatalogPage/>}/>
+        <Route path="/shop" />
+        <Route path="/user-agreement" />
+        <Route path="/cart" />
+        <Route path="/order" />
+        <Route path="*" />
+      </Routes>
+      <Footer />
+    </>
   );
 };
