+++ src/Features/slices/NavigationId.slice.js
@@ -1,0 +1,16 @@
+import {createSlice} from "@reduxjs/toolkit"
+
+export const NavigationIdSlice = createSlice({
+    name:"Navigation",
+    initialState: "",
+    reducers: {
+        setNavItem: (state, {payload}) => {
+            return payload
+       
+        }
+    }
+})
+
+
+
+export const {actions, reducer} = NavigationIdSlice
