+++ src/pages/CatalogPage/CatalogPage.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import "./CatalogPage.scss"
+import { NavigationCatalog } from '../../Components/NavigationCatalog/NavigationCatalog'
+const CatalogPage = () => {
+  return (
+    <>
+        <NavigationCatalog/>
+    </>
+  )
+}
+
+export default CatalogPage
