+++ src/pages/CatalogPage/CatalogPage.scss
@@ -1,0 +1,1 @@
+@import "../../styles/variables.scss";
